fix(fe/find-answer): Ensure every text sticker has an associated renderer_ref
