+++ frontend/apps/crates/components/src/stickers/text/state.rs
@@ -1,4 +1,8 @@
 use futures_signals::signal::Mutable;
+use js_sys::Reflect;
+use utils::unwrap::UnwrapJiExt;
+use wasm_bindgen::prelude::*;
+use web_sys::HtmlElement;
 
 use std::rc::Rc;
 
@@ -15,6 +19,8 @@ pub struct Text {
     pub value: Mutable<String>,
     pub transform: Rc<TransformState>,
     pub editor: Rc<TextEditor>,
+    /// Optional reference to the wysiwyg-output-renderer
+    pub renderer_ref: Mutable<Option<HtmlElement>>,
     pub is_editing: Mutable<bool>,
 }
 
@@ -51,6 +57,7 @@ impl Text {
                 transform_callbacks,
             )),
             editor,
+            renderer_ref: Mutable::new(None),
             is_editing,
         }
     }
@@ -60,5 +67,19 @@ impl Text {
             value: self.value.get_cloned(),
             transform: self.transform.get_inner_clone(),
         }
+    }
+
+    /// Retrieves the text value without any formatting or tags
+    pub fn get_text_value(&self) -> Option<String> {
+        let renderer_ref = &*self.renderer_ref.lock_ref();
+        renderer_ref
+            .clone()
+            .map(|renderer_ref| {
+                let value =
+                    Reflect::get(&renderer_ref, &JsValue::from_str("textValue")).unwrap_ji();
+
+                value.as_string()
+            })
+            .unwrap_or_default()
     }
 }
