+++ frontend/apps/crates/components/src/text_editor/state.rs
@@ -24,8 +24,6 @@ use super::wysiwyg_types::{
 pub struct TextEditor {
     pub controls: Mutable<ControlsState>,
     pub wysiwyg_ref: Mutable<Option<HtmlElement>>,
-    /// Optional reference to the wysiwyg-output-renderer
-    pub renderer_ref: Mutable<Option<HtmlElement>>,
     pub fonts: Mutable<Vec<String>>,
     pub value: RefCell<Option<String>>,
     pub theme_id: ReadOnlyMutable<ThemeId>,
@@ -44,7 +42,6 @@ impl TextEditor {
         let _self = Rc::new(Self {
             controls: Mutable::new(ControlsState::new()),
             wysiwyg_ref: Mutable::new(None),
-            renderer_ref: Mutable::new(None),
             fonts: Mutable::new(vec![]),
             callbacks,
             value: RefCell::new(value),
@@ -92,20 +89,6 @@ impl TextEditor {
                 let _ = reset_value_method.call0(wysiwyg_ref);
             }
         };
-    }
-
-    /// Retrieves the text value without any formatting or tags
-    pub fn get_text_value(&self) -> Option<String> {
-        let renderer_ref = &*self.renderer_ref.lock_ref();
-        renderer_ref
-            .clone()
-            .map(|renderer_ref| {
-                let value =
-                    Reflect::get(&renderer_ref, &JsValue::from_str("textValue")).unwrap_ji();
-
-                value.as_string()
-            })
-            .unwrap_or_default()
     }
 
     fn handle_fonts(state: Rc<TextEditor>) {
